Proof of concept map save/load

+++ src/game/server/entities/character.h
@@ -257,7 +257,7 @@ public:
 	//usefull everywhere
 	void DDPP_TakeDamageInstagib(int Dmg, int From, int Weapon);
 	void MoveTee(int x, int y);
-	void ChillTelePort(int X, int Y);
+	void ChillTelePort(float X, float Y);
 	void ChillTelePortTile(int X, int Y);
 	void FreezeAll(int seconds);
 	bool HasWeapon(int weapon);
